Refactor galaxy_install.tools into galaxy.
