+++ lib/galaxy/tool_shed/galaxy_install/tools/data_manager.py
