+++ lib/galaxy/tool_shed/galaxy_install/tools/tool_panel_manager.py
