+++ lib/tool_shed/galaxy_install/repair_repository_manager.py
@@ -1,10 +1,10 @@
 import logging
 import tempfile
 
+from galaxy.tool_shed.galaxy_install.tools import tool_panel_manager
 from galaxy.util import unicodify
 from tool_shed.galaxy_install import install_manager
 from tool_shed.galaxy_install.repository_dependencies import repository_dependency_manager
-from tool_shed.galaxy_install.tools import tool_panel_manager
 from tool_shed.util import basic_util
 from tool_shed.util import common_util
 from tool_shed.util import container_util
